refactor(chart): extract formatTrendAxisTick to shared chart-utils module
Move formatTrendAxisTick from dashboard.ts into
frontend/src/modules/chart-utils.ts so it can be reused by the
Purchases Savings History chart. Re-export from dashboard.ts for
backward compatibility with existing tests and callers.
